Fix: Names of Tcl/Tk resources on Tkinter hooks.
The Pyinstaller Tkinter hook was trying to copy the Tcl binary from
"/Library/Frameworks/Tcl.framework/Versions/8.5" to the dist directory,
and then the "Resources/Scripts" folder as "tcl" in that same directory,
when building an OS X Tcl/Tk application. Mac OS X filesystems are
usually case insensitive, so the folder creation was failing. The same
was happening to the Tk binary and resource folder.

Renaming the Tcl/Tk resource folders on the build hook to "TclResources"
and "TkResources" appear to fix the problem. The environment variables
in the runtime hook were updated accordingly.

See https://github.com/pyinstaller/pyinstaller/issues/1737.